Move IsInitialBlockDownload from main to checkpoints
Rebased-By: Wladimir J. van der Laan <laanwj@gmail.com>
