+++ src/main.h
@@ -12,6 +12,7 @@
 
 #include "bignum.h"
 #include "chainparams.h"
+#include "checkpoints.h"
 #include "coins.h"
 #include "core.h"
 #include "net.h"
@@ -159,8 +160,6 @@ bool CheckProofOfWork(uint256 hash, unsigned int nBits);
 unsigned int ComputeMinWork(unsigned int nBase, int64_t nTime);
 /** Get the number of active peers */
 int GetNumBlocksOfPeers();
-/** Check whether we are doing an initial block download (synchronizing from disk or network) */
-bool IsInitialBlockDownload();
 /** Format a string that describes several potential problems detected by the core */
 std::string GetWarnings(std::string strFor);
 /** Retrieve a transaction (from memory pool, or from disk, if possible) */
@@ -357,7 +356,7 @@ public:
 
         // Flush stdio buffers and commit to disk before returning
         fflush(fileout);
-        if (!IsInitialBlockDownload())
+        if (!Checkpoints::IsInitialBlockDownload())
             FileCommit(fileout);
 
         return true;
