+++ src/qt/clientmodel.cpp
@@ -133,7 +133,7 @@ QString ClientModel::getNetworkName() const
 
 bool ClientModel::inInitialBlockDownload() const
 {
-    return IsInitialBlockDownload();
+    return Checkpoints::IsInitialBlockDownload();
 }
 
 enum BlockSource ClientModel::getBlockSource() const
